+++ Runtime/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 using System.Reflection;
 
-[assembly: AssemblyVersion("0.33.0")]
+[assembly: AssemblyVersion("0.35.0")]