+++ Runtime/UI/LoginOptionsDialogPresenter.cs
@@ -8,6 +8,10 @@ namespace com.noctuagames.sdk.UI
     internal class LoginOptionsDialogPresenter : Presenter<AuthenticationModel>
     {
         private readonly ILogger _log = new NoctuaLogger();
+
+        // Flags
+        private bool _ssoDisabled = false;
+
         private Label _tnCLabel;
         private Label _privacyLabel;
         private Button _loginWithGoogleButton;
@@ -17,9 +21,15 @@ namespace com.noctuagames.sdk.UI
         private Button _registerButton;
         private Button _backButton;
         private GlobalConfig _config;
+
         public void Show()
         {
-            Visible = true;
+            if (_ssoDisabled)
+            {
+                Model.ShowEmailLogin(null);
+            } else {
+                Visible = true;
+            }
         }
 
         protected override void Attach(){}
@@ -55,6 +65,11 @@ namespace com.noctuagames.sdk.UI
         public void SetWhitelabel(GlobalConfig config)
         {
             _config = config;
+        }
+
+        public void SetFlag(bool SSODisabled = false)
+        {
+            _ssoDisabled = SSODisabled;
         }
 
         private void OnTnCClicked()